// bamul/src/app/shared/model/driver.ts
export class Driver {
    driverName: string;
    gender: string;
    driverDateofBirth: Date;
    mobileNumber: string;
    licenceNumber: string;
    status: string;
    licenceExpDate: Date;
    state: string;
    city: string;
    nomineeName: string;
    nomineeRelation: string;
    nomineeDob: Date;
    licenceType: string;
    address: string;
    country: string;
    driverId: number;
    driverPhoto: Blob;
    creationDate:Date;
    createdBy:number;
   lastUpdateDate:Date;
   lastUpdatedBy:number;
}